+++ packages/slice-machine/src/modules/screenshots/actions.ts
@@ -1,0 +1,34 @@
+import { ActionType, createAsyncAction } from "typesafe-actions";
+import { Screenshots } from "@lib/models/common/Screenshots";
+import { ComponentUI, ScreenshotUI } from "@lib/models/common/ComponentUI";
+
+export type SelectedSliceActions =
+  | ActionType<typeof generateSliceScreenshotCreator>
+  | ActionType<typeof generateSliceCustomScreenshotCreator>;
+
+export const generateSliceScreenshotCreator = createAsyncAction(
+  "SLICE/TAKE_SCREENSHOT.REQUEST",
+  "SLICE/TAKE_SCREENSHOT.RESPONSE",
+  "SLICE/TAKE_SCREENSHOT.FAILURE"
+)<
+  {
+    variationId: string;
+    component: ComponentUI;
+    setData: (data: any) => void;
+  },
+  { screenshots: Screenshots; component: ComponentUI }
+>();
+
+export const generateSliceCustomScreenshotCreator = createAsyncAction(
+  "SLICE/GENERATE_CUSTOM_SCREENSHOT.REQUEST",
+  "SLICE/GENERATE_CUSTOM_SCREENSHOT.RESPONSE",
+  "SLICE/GENERATE_CUSTOM_SCREENSHOT.FAILURE"
+)<
+  {
+    variationId: string;
+    component: ComponentUI;
+    setData: (data: any) => void;
+    file: Blob;
+  },
+  { variationId: string; screenshot: ScreenshotUI; component: ComponentUI }
+>();
