+++ packages/slice-machine/src/modules/selectedSlice/actions.ts
@@ -2,8 +2,7 @@ import { ActionType, createAction, createAsyncAction } from "typesafe-actions";
 import { Models } from "@slicemachine/core";
 import { NestableWidget } from "@prismicio/types-internal/lib/customtypes/widgets/nestable";
 import { SliceMockConfig } from "@lib/models/common/MockConfig";
-import { Screenshots } from "@lib/models/common/Screenshots";
-import { ComponentUI, ScreenshotUI } from "@lib/models/common/ComponentUI";
+import { ComponentUI } from "@lib/models/common/ComponentUI";
 import { renameSliceCreator } from "../slices";
 import { SelectedSliceStoreType } from "./types";
 import { SliceBuilderState } from "../../../lib/builders/SliceBuilder";
@@ -17,8 +16,6 @@ export type SelectedSliceActions =
   | ActionType<typeof removeSliceWidgetCreator>
   | ActionType<typeof updateSliceWidgetMockCreator>
   | ActionType<typeof deleteSliceWidgetMockCreator>
-  | ActionType<typeof generateSliceScreenshotCreator>
-  | ActionType<typeof generateSliceCustomScreenshotCreator>
   | ActionType<typeof saveSliceCreator>
   | ActionType<typeof pushSliceCreator>
   | ActionType<typeof copyVariationSliceCreator>
@@ -74,33 +71,6 @@ export const deleteSliceWidgetMockCreator = createAction(
   widgetArea: Models.WidgetsArea;
   newKey: string;
 }>();
-
-export const generateSliceScreenshotCreator = createAsyncAction(
-  "SLICE/TAKE_SCREENSHOT.REQUEST",
-  "SLICE/TAKE_SCREENSHOT.RESPONSE",
-  "SLICE/TAKE_SCREENSHOT.FAILURE"
-)<
-  {
-    _variationId: string;
-    component: ComponentUI;
-    setData: (data: any) => void;
-  },
-  { screenshots: Screenshots; component: ComponentUI }
->();
-
-export const generateSliceCustomScreenshotCreator = createAsyncAction(
-  "SLICE/GENERATE_CUSTOM_SCREENSHOT.REQUEST",
-  "SLICE/GENERATE_CUSTOM_SCREENSHOT.RESPONSE",
-  "SLICE/GENERATE_CUSTOM_SCREENSHOT.FAILURE"
-)<
-  {
-    variationId: string;
-    component: ComponentUI;
-    setData: (data: any) => void;
-    file: Blob;
-  },
-  { variationId: string; screenshot: ScreenshotUI; component: ComponentUI }
->();
 
 export const saveSliceCreator = createAsyncAction(
   "SLICE/SAVE.REQUEST",
