+++ packages/slice-machine/tests/src/modules/screenshots/__mockData__/model.json
@@ -1,0 +1,35 @@
+{
+  "id": "dummy_slice",
+  "type": "SharedSlice",
+  "name": "DummySlice",
+  "description": "DummySlice",
+  "variations": [
+    {
+      "id": "default-slice",
+      "name": "Default slice",
+      "docURL": "...",
+      "version": "sktwi1xtmkfgx8626",
+      "description": "DummySlice",
+      "primary": {
+        "section_title": {
+          "config": {
+            "label": "Section Title",
+            "placeholder": ""
+          },
+          "type": "Text"
+        }
+      },
+      "items": {
+        "link": {
+          "config": {
+            "label": "Link",
+            "placeholder": "",
+            "select": "document",
+            "allowTargetBlank": false
+          },
+          "type": "Link"
+        }
+      }
+    }
+  ]
+}
