+++ packages/slice-machine/tests/src/modules/selectedSlice/sagas.test.ts
@@ -1,17 +1,11 @@
 import { testSaga } from "redux-saga-test-plan";
-import {
-  pushSliceSaga,
-  saveSliceSaga,
-  generateSliceScreenshotSaga,
-} from "@src/modules/selectedSlice/sagas";
+import { pushSliceSaga, saveSliceSaga } from "@src/modules/selectedSlice/sagas";
 import { getSelectedSliceDummyData } from "./utils";
 import {
-  generateSliceScreenshotCreator,
   pushSliceCreator,
   saveSliceCreator,
 } from "../../../../src/modules/selectedSlice/actions";
 import {
-  generateSliceScreenshotApiClient,
   pushSliceApiClient,
   saveSliceApiClient,
 } from "../../../../src/apiClient";
@@ -22,7 +16,7 @@ import {
 import { getRemoteSlice } from "@src/modules/slices";
 import { SliceSM } from "@slicemachine/core/build/models";
 
-const { dummySliceState, dummyModelVariationID } = getSelectedSliceDummyData();
+const { dummySliceState } = getSelectedSliceDummyData();
 
 describe("[Selected Slice sagas]", () => {
   describe("[saveSliceSaga]", () => {
@@ -126,78 +120,6 @@ describe("[Selected Slice sagas]", () => {
       saga.throw(new Error()).put(
         openToasterCreator({
           message: "Internal Error: Slice was not pushed",
-          type: ToasterType.ERROR,
-        })
-      );
-      saga.next().isDone();
-    });
-  });
-
-  describe("[generateSliceScreenshotSaga]", () => {
-    it("should call the api and dispatch the success action", () => {
-      const mockSetData = jest.fn();
-      const saga = testSaga(
-        generateSliceScreenshotSaga,
-        generateSliceScreenshotCreator.request({
-          _variationId: dummyModelVariationID,
-          component: dummySliceState,
-          setData: mockSetData,
-        })
-      );
-
-      saga
-        .next()
-        .call(
-          generateSliceScreenshotApiClient,
-          dummySliceState.model.name,
-          dummySliceState.from
-        );
-      const response = {
-        screenshots: {
-          dummyModelVariationID: {
-            path: "testScreenshotPath",
-            url: "testScreenshotUrl",
-          },
-        },
-      };
-
-      saga
-        .next({
-          status: 200,
-          data: response,
-        })
-        .put(
-          generateSliceScreenshotCreator.success({
-            screenshots: response.screenshots,
-            component: dummySliceState,
-          })
-        );
-
-      saga.next().isDone();
-      expect(mockSetData).toHaveBeenCalledWith({
-        done: true,
-        error: null,
-        loading: false,
-        imageLoading: false,
-        message: "Screenshots were saved to FileSystem",
-        status: 200,
-        warning: false,
-      });
-    });
-    it("should open a error toaster on internal error", () => {
-      const mockSetData = jest.fn();
-      const saga = testSaga(
-        generateSliceScreenshotSaga,
-        generateSliceScreenshotCreator.request({
-          _variationId: dummyModelVariationID,
-          component: dummySliceState,
-          setData: mockSetData,
-        })
-      ).next();
-
-      saga.throw(new Error()).put(
-        openToasterCreator({
-          message: "Internal Error: Screenshot not saved",
           type: ToasterType.ERROR,
         })
       );
